pyproject.toml
[project]
name = "video-downloader-pro"
version = "0.1.0"
description = "Download and process videos/images from social media - remove watermark and fingerprints"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.12.0",
    "opencv-python>=4.9.0",
    "ffmpeg-python>=0.2.0",
    "numpy>=1.26.0",
    "Pillow>=10.2.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "instaloader>=4.10.5",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
ocr = [
    "easyocr>=1.7.1",
]

[project.scripts]
vdl-pro = "cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["downloader", "processor", "utils", "cli.py"]
